requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
Flask>=3.0.0
PyYAML>=6.0.0
paramiko>=3.4.0
scp>=0.15.0
